backend/routes/BFSI-LOS-Model_Pdf/requirements.txt
pytesseract
Pillow
openai
python-dotenv
pandas
numpy
scikit-learn
argparse
yfinance
openpyxl
PyPDF2
pymongo